% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_images.R
\name{as.otolith_contour_list}
\alias{as.otolith_contour_list}
\title{Convert to otolith_contour_list}
\usage{
as.otolith_contour_list(x, ...)
}
\arguments{
\item{x}{otolith_contour object or output from read_image}

\item{...}{other parameters}
}
\value{
otolith_contour_list object
}
\description{
Convert to otolith_contour_list
}
\author{
Christoffer Moesgaard Albertsen
}
