% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_unknown.R
\name{assign_unknown}
\alias{assign_unknown}
\title{Assign unknown contours by binomial regression on elliptic fourier descriptors}
\usage{
assign_unknown(nimList, N = 60, delta = 0.01, lambda = 1,
  link = "probit")
}
\arguments{
\item{nimList}{A list of normalized contours}

\item{N}{Number of harmonics for the fourier descriptors}

\item{delta}{passed to fcbf for feature selection}

\item{lambda}{passed to fcbf for feature selection}
}
\value{
A list of normalized contours where 'Unknown' Position attributes are replaced by 'Left' or 'Right'.
}
\description{
Assign unknown contours by binomial regression on elliptic fourier descriptors
}
\author{
Christoffer Moesgaard Albertsen
}
