% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conte.R
\name{Conte}
\alias{Conte}
\title{From Morphometrics with R
Modified to start at right most non-zero pixel}
\usage{
Conte(imagematrix)
}
\description{
From Morphometrics with R
Modified to start at right most non-zero pixel
}
