% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim.R
\name{discrim}
\alias{discrim}
\title{Discriminant analysis}
\usage{
discrim(train, group, test, type = "lda", verbose = TRUE, dist = "normal",
  prior = table(group)/length(group), nrr = rank)
}
\arguments{
\item{train}{A matrix of (continuous) features for the training sample}

\item{group}{A vector of group labels for the training sample}

\item{test}{A matrix of (continuous) features for the test sample}

\item{type}{Type of discriminant analysis to use. lda for linear and qda for quadratic.}

\item{verbose}{Write maximum gradient components to the terminal?}

\item{dist}{Distribution to use (currently ignored)}
}
\value{
stuff
}
\description{
\code{discrim} performs linear and quadratic discriminant analysis.
}
