% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim.R
\name{discrimTMB}
\alias{discrimTMB}
\title{discrim with TMB}
\usage{
discrimTMB(train, group, test, type = "lda", verbose = TRUE,
  dist = "normal", prior = table(group)/length(group))
}
\arguments{
\item{train}{training data}

\item{group}{training groups}

\item{test}{test data}

\item{type}{type of analysis: lda or qda}

\item{verbose}{should a lot of output be printed?}

\item{dist}{not currently used}

\item{prior}{prior probability for groups}
}
\value{
...
}
\description{
discrim with TMB
}
\author{
Christoffer Moesgaard Albertsen
}
