% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efd.R
\name{efd2coord}
\alias{efd2coord}
\title{Calculate XY-coordinates from Elliptical Fourier Descriptors}
\usage{
efd2coord(n, A, B, C, D, A0 = 0, C0 = 0)
}
\arguments{
\item{n}{}

\item{A}{}

\item{B}{}

\item{C}{}

\item{D}{}

\item{A0}{}

\item{C0}{}
}
\value{
Matrix of coordinates
}
\description{
Calculate XY-coordinates from Elliptical Fourier Descriptors
}
\author{
Christoffer Moesgaard Albertsen
}
