% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efd.R
\name{efd2coordSEXP}
\alias{efd2coordSEXP}
\title{Calculate XY-coordinates from Elliptical Fourier Descriptors}
\usage{
efd2coordSEXP(n, A, B, C, D, A0 = 0, C0 = 0)
}
\arguments{
\item{n}{Number of coordinates to calculate}

\item{A}{A EFD coefficients}

\item{B}{B EFD coefficients}

\item{C}{C EFD coefficients}

\item{D}{D EFD coefficients}

\item{A0}{A0 EFD coefficients}

\item{C0}{C0 EFD coefficients}
}
\value{
Matrix of coordinates
}
\description{
Calculate XY-coordinates from Elliptical Fourier Descriptors
}
\author{
Christoffer Moesgaard Albertsen
}
