% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbf.R
\name{fcbf}
\alias{fcbf}
\title{Fast Correlation Based Filter for Feature Selection in High-Dimensional Data}
\usage{
fcbf(train, group, delta, lambda = 1)
}
\arguments{
\item{train}{A matrix of training data}

\item{group}{A vector of groups}

\item{delta}{Tuning parameter}

\item{lambda}{Tuning parameter}
}
\value{
A vector of selected features
}
\description{
Fast Correlation Based Filter for Feature Selection in High-Dimensional Data
}
\references{
http://pdf.aminer.org/000/335/746/feature_selection_for_high_dimensional_data_a_fast_correlation_based.pdf
}
\author{
Christoffer Moesgaard Albertsen
}
