% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip an otolith}
\usage{
flip(x, axis, ...)
}
\arguments{
\item{x}{otolith object}

\item{axis}{axis to flip around}

\item{...}{not used}
}
\value{
a flipped otolith object
}
\description{
Flip an otolith
}
\author{
Christoffer Moesgaard Albertsen
}
