% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efd.R
\name{fs}
\alias{fs}
\title{Fourier Series Basis for Splines}
\usage{
fs(x, N, Boundary = range(x))
}
\arguments{
\item{x}{the predictor variable}

\item{N}{Half the number of basis functions. Both a cos and a sin basis function will be added.}

\item{Boundary}{Boundary points that determines a period}
}
\value{
A matrix of dimension 'c(length(x), 2*N)'.
}
\description{
Fourier Series Basis for Splines
}
\author{
Christoffer Moesgaard Albertsen
}
