% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getGroupMeans}
\alias{getGroupMeans}
\alias{getGroupMeans.mlld}
\title{Calculate group means for a data frame of covariates}
\usage{
getGroupMeans(f, data, ...)

\method{getGroupMeans}{mlld}(f, data, keep.cov = FALSE, ...)
}
\arguments{
\item{f}{}

\item{data}{}
}
\value{

}
\description{
Calculate group means for a data frame of covariates
}
\author{
Christoffer Moesgaard Albertsen
}
