% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getGroupProportion}
\alias{getGroupProportion}
\alias{getGroupProportion.mlld}
\title{Calculate group proportions for a data frame of covariates}
\usage{
getGroupProportion(f, data)

\method{getGroupProportion}{mlld}(f, data, randEff = TRUE)
}
\arguments{
\item{f}{}

\item{data}{}
}
\value{

}
\description{
Calculate group proportions for a data frame of covariates
}
\author{
Christoffer Moesgaard Albertsen
}
