% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{getPixelMatrix}
\alias{getPixelMatrix}
\title{Get a pixel matrix from an image file}
\usage{
getPixelMatrix(file, grey = TRUE)
}
\arguments{
\item{file}{Path to image file}

\item{grey}{Should output be greyscale?}
}
\value{
A matrix of pixel values (0-255)
}
\description{
Get a pixel matrix from an image file
}
\author{
Christoffer Moesgaard Albertsen
}
