% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate otolith coordinates through EFDs}
\usage{
interpolate(x, N, Nefd, ...)
}
\arguments{
\item{x}{otolith object}

\item{N}{Number of coordinates to output}

\item{Nefd}{Number of EFDs to use}

\item{...}{not used}
}
\value{
an interpolated otolith object
}
\description{
Interpolate otolith coordinates through EFDs
}
\author{
Christoffer Moesgaard Albertsen
}
