% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{k Nearest Neighbour}
\usage{
knn(train, group, test, kn = 1, dist = c("L2", "L1", "Linf"))
}
\arguments{
\item{train}{A matrix of (continuous) features for the training sample}

\item{group}{A vector of group labels for the training sample}

\item{test}{A matrix of (continuous) features for the test sample}

\item{kn}{Number of neighbours to consider}

\item{dist}{Distance measure to use ("L1", "L2","Linf"). Default is Euclidian distance ("L2").}
}
\description{
\code{knn} returns the votes to each level factor(group) from the kn nearest neighbours to test from train.
}
