% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcd.R
\name{mlcd}
\alias{mlcd}
\title{Maximum Likelihood Contour Discrimination}
\usage{
mlcd(
  train,
  group,
  test,
  prior = as.vector(table(group))/length(group),
  penalty = 0,
  lambda = 0.4,
  Nefd = 60,
  correlatedCoordinates = FALSE,
  equalVariance = TRUE,
  silent = FALSE,
  control = list(iter.max = 1e+05, eval.max = 1e+05),
  ...
)
}
\arguments{
\item{train}{3D Array of training data (2 x number of contour points x number of observations)}

\item{group}{Factor of training groups}

\item{test}{3D array of test data}

\item{prior}{Prior probability of groups}

\item{penalty}{p to use for Lp penalty. Zero is no penalty.}

\item{lambda}{Positive scalar factor for Lp penalty. Zero is no penalty.}

\item{Nefd}{Number of EFDs to use for mean otolith}

\item{correlatedCoordinates}{Should coordinates be correlated?}

\item{equalVariance}{Should the variance be equal in the two coordinates?}

\item{silent}{Should the TMB object be silent?}

\item{control}{control parameters passes to nlminb}

\item{...}{Other parameters}
}
\value{
a list of the result
}
\description{
Maximum Likelihood Contour Discrimination
}
\author{
Christoffer Moesgaard Albertsen
}
