% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlld.R
\name{mlld}
\alias{mlld}
\title{Maximum Likelihood Linear Discrimination}
\usage{
mlld(train, group, test, prior = as.vector(table(group))/length(group),
  featureSelection = c("none", "fcbf", "Lp"), penalty = 0,
  lambda = 0.4, estimateLambda = TRUE, independent = FALSE,
  silent = FALSE, control = list(iter.max = 1e+05, eval.max = 1e+05),
  formula = ~1, formulaCommon = ~1, data = NULL, dataTest = NULL,
  estimateUnbiasedTestProportions = FALSE,
  trainContributionToTestProportions = FALSE,
  biasCorrectionGroup = NULL, trainBiasCorrectionGroup = NULL,
  addMisfitCategory = FALSE, fcbfDelta = 0.01, fcbfLambda = 1,
  onlyObj = FALSE, doSdreport = FALSE, ...)
}
\arguments{
\item{train}{Matrix of training data (number of observations x number of features)}

\item{group}{Factor of training groyps}

\item{test}{Matrix of test data}

\item{prior}{Prior probability of groups}

\item{penalty}{p to use for Lp penalty. Zero is no penalty}

\item{lambda}{Positive numeric vector factor for Lp penalty. Zero is no penalty. The first element determines the penalty on group differences, the second element determines the penalty on absolute parameter values, and the third element determines the penalty on correlation parameters.}

\item{independent}{Should features be treated as independent?}

\item{silent}{Should the TMB object be silent?}

\item{control}{control parameters passes to nlminb}

\item{formula}{Formula to account for in the linear discrimination}

\item{data}{Covariates for training data}

\item{dataTest}{Covariates for testing data}

\item{...}{Other parameters}
}
\value{
a list of the result
}
\description{
Equivalent to LDA
}
\author{
Christoffer Moesgaard Albertsen
}
