% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{normalize_image}
\alias{normalize_image}
\title{Normalize an Otolith Image}
\usage{
normalize_image(dat, n, datCompare = NULL, forceFlip = FALSE,
  flipByPosition = c("No", "Left", "Right"))
}
\arguments{
\item{dat}{}

\item{n}{}

\item{datCompare}{}

\item{forceFlip}{}

\item{flipByPosition}{}
}
\value{
A normalized Otolith Image
}
\description{
Normalize an Otolith Image
}
\author{
Christoffer Moesgaard Albertsen
}
