\name{otoclass-package}
\alias{otoclass-package}
\alias{otoclass}
\docType{package}
\title{
Otolith Classification and Proportion Estimation
}
\description{
This packages provides classification methods for analysis of
    otoliths along with methods for estimating species proportions in samples.
}
\details{
\tabular{ll}{
Package: \tab otoclass\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-12\cr
License: \tab GPL-2\cr
}

}
\author{
Christoffer Moesgaard Albertsen <cmoe@aqua.dtu.dk>
}
\references{
}
\keyword{ otolith, classification }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
