% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_methods.R
\name{plot.otolith_contour}
\alias{plot.otolith_contour}
\title{Plot otolith contour}
\usage{
\method{plot}{otolith_contour}(x, asUsed = FALSE, ...)
}
\arguments{
\item{x}{Otolith contour}

\item{asUsed}{Plot image as read; i.e. transformed}

\item{...}{further arguments}
}
\value{
nothing
}
\description{
Plot otolith contour
}
\author{
Christoffer Moesgaard Albertsen
}
