% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_methods.R
\name{plot.otolith_image}
\alias{plot.otolith_image}
\title{Plot otolith image}
\usage{
\method{plot}{otolith_image}(x, asUsed = FALSE, ...)
}
\arguments{
\item{x}{Otolith image}

\item{asUsed}{Plot image as read; i.e. transformed}

\item{...}{further arguments}
}
\value{
nothing
}
\description{
Plot otolith image
}
\author{
Christoffer Moesgaard Albertsen
}
