% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon_area}
\alias{polygon_area}
\title{Signed area of a closed polygon}
\usage{
polygon_area(X)
}
\arguments{
\item{X}{A numeric matrix with two columns of vertices}
}
\value{
Polygon area in same units as input
}
\description{
The function uses the shoelace formula to calculate the signed area of a closed non-intersecting polygon.
}
\author{
Christoffer Moesgaard Albertsen
}
