% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon_centroid}
\alias{polygon_centroid}
\title{Centroid of a closed polygon}
\usage{
polygon_centroid(X)
}
\arguments{
\item{X}{A numeric matrix with two columns of vertices}
}
\value{
Polygon centroid in same units as input
}
\description{
The function calculates the centroid of a closed polygon.
}
\author{
Christoffer Moesgaard Albertsen
}
