% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_methods.R
\name{print.otolith_contour}
\alias{print.otolith_contour}
\title{Print otolith image contour}
\usage{
\method{print}{otolith_contour}(x, ...)
}
\arguments{
\item{x}{Otolith image contour}

\item{...}{further arguments}
}
\value{
nothing
}
\description{
Print otolith image contour
}
\author{
Christoffer Moesgaard Albertsen
}
