% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_methods.R
\name{print.otolith_image}
\alias{print.otolith_image}
\title{Print otolith image}
\usage{
\method{print}{otolith_image}(x, ...)
}
\arguments{
\item{x}{Otolith image}

\item{...}{further arguments}
}
\value{
nothing
}
\description{
Print otolith image
}
\author{
Christoffer Moesgaard Albertsen
}
