% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projection}
\alias{projection}
\title{LDA-like projection}
\usage{
projection(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
LDA-like projection
}
\author{
Christoffer Moesgaard Albertsen
}
