% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{Read Otolith Images and Extract Contours}
\usage{
read_image(file, noiseFactor = NULL, onlyOne = FALSE, minPixelDiff = 0.05
  * min(nc, nr), extreme = TRUE, borderBasedCutOff = FALSE,
  logisticTransform = FALSE, logisticTransformLocation = c("mean", "median",
  "borderMean", "borderMedian"), logisticTransformScale = 1,
  gaussianBlur = FALSE, gaussianBlurSize = round(0.02 * min(nc, nr)),
  gaussianBlurSigma = gaussianBlurSize/10, unsharp = FALSE,
  pixelwise = FALSE, assignSinglesByPosition = TRUE, minCountScale = 0,
  minCountForMax = 1e-04, zeroCutOffPercent = 0.05)
}
\arguments{
\item{file}{Image file path}

\item{noiseFactor}{Scalar value determining cutoff value for background noise on image. If NULL, a value is calculated automatically.}

\item{onlyOne}{Boolean value. Is there only one otolith on the image?}

\item{minPixelDiff}{Minimum pixel difference between otoliths}

\item{extreme}{Boolean value. Should pixel values be converted to 0/1?}

\item{pixelwise}{Boolean value. If TRUE, a pixel-wise algorithm is used to extract contours; otherwise, \code{grDevices::contourLines} is used.}

\item{assignSinglesByPosition}{Should single otoliths be assigned to Left/Right based on position on image?}

\item{minCountScale}{See details}
}
\value{
otolith image information
}
\description{
Read Otolith Images and Extract Contours
}
\author{
Christoffer Moesgaard Albertsen
}
