% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_images.R
\name{read_images}
\alias{read_images}
\title{Extract otolith contours from images}
\usage{
read_images(files, ...)
}
\arguments{
\item{files}{file paths to read}

\item{...}{Arguments passed to read_image}
}
\value{
an otolith_contour_list object
}
\description{
Extract otolith contours from images
}
\seealso{
read_image
}
\author{
Christoffer Moesgaard Albertsen
}
