% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate an otolith}
\usage{
rotate(x, angle, degrees = TRUE, ...)
}
\arguments{
\item{x}{object to rotate}

\item{angle}{Angles to rotate the object}

\item{degrees}{Is the angle in degrees?}

\item{...}{not used}
}
\value{
A rotated element
}
\description{
Rotate an otolith
}
\author{
Christoffer Moesgaard Albertsen
}
