% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyExample.R
\name{runShinyExample}
\alias{runShinyExample}
\title{Run shiny example included in the package}
\usage{
runShinyExample(name)
}
\arguments{
\item{name}{Name of the shiny example (biascorrect, or proportion)}
}
\value{
Nothing
}
\description{
Run shiny example included in the package
}
\author{
Christoffer Moesgaard Albertsen
}
