% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim.R
\name{discrimTMB}
\alias{discrimTMB}
\title{discrim with TMB}
\usage{
discrimTMB(train, group, test, type = "lda", verbose = TRUE,
  dist = "normal", prior = table(group)/length(group))
}
\arguments{
\item{train}{}

\item{group}{}

\item{test}{}

\item{type}{}

\item{verbose}{}

\item{dist}{}

\item{prior}{}
}
\value{
...
}
\description{
discrim with TMB
}
\author{
Christoffer Moesgaard Albertsen
}
