% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efd.R
\name{efd}
\alias{efd}
\title{Calculate Elliptical Fourier Descriptors}
\usage{
efd(dat, N, n = nrow(dat), returnAsList = FALSE)
}
\arguments{
\item{dat}{}

\item{N}{}

\item{n}{}

\item{returnAsList}{}
}
\value{
List of elliptical fourier descriptors
}
\description{
Calculate Elliptical Fourier Descriptors
}
\author{
Christoffer Moesgaard Albertsen
}
