% Generated by roxygen2 (4.0.2): do not edit by hand
\name{knn}
\alias{knn}
\title{k Nearest Neighbour}
\usage{
knn(train, group, test, kn, dist = "euclidian")
}
\arguments{
\item{train}{A matrix of (continuous) features for the training sample}

\item{group}{A vector of group labels for the training sample}

\item{test}{A matrix of (continuous) features for the test sample}

\item{kn}{Number of neighbours to consider}

\item{dist}{Distance measure to use ("euclidian", "manhattan","chebyshev")}
}
\description{
\code{knn} returns the votes to each level factor(group) from the kn nearest neighbours to test from train.
}

