% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlld.R
\name{mlld}
\alias{mlld}
\title{Maximum likelihood linear discrimination}
\usage{
mlld(
  y,
  group,
  data = NULL,
  formula = ~1,
  formulaCommon = ~1,
  formulaProportion = ~1,
  formulaLogScale = ~-1,
  equalLogScale = FALSE,
  lp_penalty = NA,
  lambda = 0.4,
  estimateLambda = TRUE,
  REML = FALSE,
  tMixture = 0,
  tDf = 5,
  estimateTMix = FALSE,
  sameTMix = FALSE,
  estimateTDf = FALSE,
  sameTDf = FALSE,
  independent = FALSE,
  silent = FALSE,
  control = list(iter.max = 1e+05, eval.max = 1e+05),
  drop.unused.levels = TRUE,
  onlyObj = FALSE,
  doSdreport = TRUE,
  getReportCovariance = TRUE,
  equalVariance = TRUE,
  confusionMatrixList = NULL,
  groupConversionList = list(),
  confusionLevelTypes = rep("Known", ifelse(is.null(ncol(group)), 1, ncol(group))),
  observationType = c("MVMIX", "SNP1", "SNP2", "SPL_AR1", "FS1_AR1", "SPL_CATEGORICAL"),
  forceMeanIncrease = FALSE,
  fixZeroGradient = TRUE,
  SNP2dm = TRUE,
  sameSNP2dm = TRUE,
  SPLknots = NULL,
  Nefd = 5,
  lower = list(),
  upper = list(),
  guessStartingValues = TRUE,
  ...
)
}
\arguments{
\item{...}{}

\item{proportionGroup}{}

\item{confusionGroup}{}
}
\value{

}
\description{
Maximum likelihood linear discrimination
}
\author{
Christoffer Moesgaard Albertsen
}
