% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlld.R
\name{mlld}
\alias{mlld}
\title{Maximum Likelihood Linear Discrimination}
\usage{
mlld(train, group, test, prior = as.vector(table(group))/length(group),
  penalty = 0, lambda = 0.4, independent = FALSE, silent = FALSE,
  control = list(iter.max = 1e+05, eval.max = 1e+05), ...)
}
\arguments{
\item{train}{Matrix of training data (number of observations x number of features)}

\item{group}{Factor of training groyps}

\item{test}{Matrix of test data}

\item{prior}{Prior probability of groups}

\item{penalty}{p to use for Lp penalty. Zero is no penalty}

\item{lambda}{Positive scalar factor for Lp penalty. Zero is no penalty.}

\item{independent}{Should features be treated as independent?}

\item{silent}{Should the TMB object be silent?}

\item{control}{control parameters passes to nlminb}

\item{...}{Other parameters}
}
\value{
a list of the result
}
\description{
Equivalent to LDA
}
\author{
Christoffer Moesgaard Albertsen
}
