% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{normalize_image}
\alias{normalize_image}
\title{Normalize an Otolith Image}
\usage{
normalize_image(
  dat,
  n,
  datCompare = NULL,
  forceFlip = FALSE,
  flipByPosition = c("No", "Left", "Right")
)
}
\arguments{
\item{dat}{otolith_image object}

\item{n}{Number of coordinates to output}

\item{datCompare}{Data to compare with}

\item{forceFlip}{Force a flip of the otolith?}

\item{flipByPosition}{Position (Left/Right) to flip. Use 'No' if otoliths should not be flipped by position.}
}
\value{
A normalized Otolith Image
}
\description{
Normalize an Otolith Image
}
\author{
Christoffer Moesgaard Albertsen
}
