% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_methods.R
\name{plot.otolith_contour_list}
\alias{plot.otolith_contour_list}
\title{Plot otolith contour list}
\usage{
\method{plot}{otolith_contour_list}(x, asUsed = FALSE, ask = TRUE, ...)
}
\arguments{
\item{x}{Otolith contour list}

\item{asUsed}{Plot image as read; i.e. transformed}

\item{ask}{Ask before plotting next contour?}

\item{...}{further arguments}
}
\value{
nothing
}
\description{
Plot otolith contour list
}
\author{
Christoffer Moesgaard Albertsen
}
