% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{Read Otolith Images and Extract Contours}
\usage{
read_image(file, noiseFactor = NULL, onlyOne = FALSE, minPixelDiff = 0.03
  * min(nc, nr), extreme = FALSE, assignSinglesByPosition = TRUE)
}
\arguments{
\item{file}{}

\item{noiseFactor}{}

\item{onlyOne}{}

\item{minPixelDiff}{}
}
\value{
otolith image information
}
\description{
Read Otolith Images and Extract Contours
}
\author{
Christoffer Moesgaard Albertsen
}
