% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{Read Otolith Images and Extract Contours}
\usage{
read_image(file, noiseFactor = NULL, onlyOne = FALSE,
  minPixelDiff = 0.05 * min(nc, nr), extreme = TRUE,
  floodFillTolerance = 0, borderBasedCutOff = FALSE,
  logisticTransform = FALSE, logisticTransformLocation = c("mean",
  "median", "borderMean", "borderMedian"), logisticTransformScale = 1,
  gaussianBlur = TRUE, gaussianBlurSize = round(0.02 * min(nc, nr)),
  gaussianBlurSigma = gaussianBlurSize/10, sobel = TRUE,
  sobelSize = 7, unsharp = FALSE, pixelwise = FALSE,
  assignSinglesByPosition = TRUE, minCountScale = 0,
  minCountForMax = 1e-04, reduceCutOffPercent = 0,
  zeroCutOffPercent = 0.05, maxOtoliths = 2, minArea = 0.001,
  cleaningTolerance = 0.2)
}
\arguments{
\item{file}{Image file path}

\item{noiseFactor}{Scalar value determining cutoff value for background noise on image. If NULL, a value is calculated automatically.}

\item{onlyOne}{Boolean value. Is there only one otolith on the image?}

\item{minPixelDiff}{Minimum pixel difference between otoliths}

\item{extreme}{Boolean value. Should pixel values be converted to 0/1?}

\item{borderBasedCutOff}{Use a border based cut off?}

\item{logisticTransform}{Do logistic transformation of image?}

\item{logisticTransformLocation}{Location parameter for logistic transformation}

\item{logisticTransformScale}{Scale parameter for logistic transformation}

\item{gaussianBlur}{Do Gaussian blur of image?}

\item{gaussianBlurSize}{Size (n x n) for Gaussian blur kernel}

\item{gaussianBlurSigma}{Variance parameter of Gaussian distribution used to blur}

\item{unsharp}{Do unsharp mask of image using the Gaussian blur arguments?}

\item{pixelwise}{Boolean value. If TRUE, a pixel-wise algorithm is used to extract contours; otherwise, \code{grDevices::contourLines} is used.}

\item{assignSinglesByPosition}{Should single otoliths be assigned to Left/Right based on position on image?}

\item{minCountScale}{See details}

\item{minCountForMax}{See details}

\item{reduceCutOffPercent}{See details}

\item{zeroCutOffPercent}{See details}
}
\value{
otolith image object
}
\description{
Read Otolith Images and Extract Contours
}
\author{
Christoffer Moesgaard Albertsen
}
